requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
cycler==0.10.0
Django==2.2.12
django-bootstrap4==1.1.1
gunicorn==20.0.4
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.3
Pillow==7.1.2
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
whitenoise==5.0.1
